Fix double-output and improves reliability
Instead of using `onload`, has the child frame post a message to the
parent on the child's DOMContentLoaded.

The examples now always run after DOMContentLoaded.

The code is now evaluated using `System.module` instead of attaching a
script, eliminating some weird error messages in the console.

Also, adds a math call to stub.rkt.

Not sure which of the above did it, but this fixes #7.